點擊信封袋新增一筆帳務

+++ AidanEnvelopeSaveMoney/app/src/main/java/com/aidan/aidanenvelopesavemoney/Model/Account.java
@@ -5,15 +5,15 @@ package com.aidan.aidanenvelopesavemoney.Model;
  */
 
 public class Account {
-    private int cost = 0;
+    private float cost = 0;
     private String comment = "";
     private String envelopeName = "";
 
-    public int getCost() {
+    public float getCost() {
         return cost;
     }
 
-    public void setCost(int cost) {
+    public void setCost(float cost) {
         this.cost = cost;
     }
